fixed warnings (#1290)
* fixed warnings


* Added cosmetic suggestions

+++ include/fastrtps/utils/fixed_size_bitmap.hpp
@@ -38,10 +38,8 @@ using std::uint32_t;
 template <class T>
 struct DiffFunction
 {
-    constexpr auto operator () (
-            T a,
-            T b) const
-    -> decltype(a - b)
+    constexpr auto operator () (T a, T b) const
+        -> decltype(a - b)
     {
         return a - b;
     }
@@ -61,7 +59,7 @@ struct DiffFunction
 template<class T, class Diff = DiffFunction<T>, uint32_t NBITS = 256>
 class BitmapRange
 {
-    #define NITEMS ((NBITS + 31ul) / 32ul)
+    #define NITEMS ((NBITS + 31u) / 32u)
 
 public:
 
@@ -184,7 +182,7 @@ public:
     {
         // Traverse through the significant items on the bitmap
         T item = base_;
-        uint32_t n_longs = (num_bits_ + 31UL) / 32UL;
+        uint32_t n_longs = (num_bits_ + 31u) / 32u;
         for (uint32_t i = 0; i < n_longs; i++)
         {
             // Check if item has at least one bit set
@@ -197,17 +195,17 @@ public:
 #if _MSC_VER
                 unsigned long bit;
                 _BitScanReverse(&bit, bits);
-                uint32_t offset = 31UL ^ bit;
+                uint32_t offset = 31u ^ bit;
 #else
-                uint32_t offset = __builtin_clz(bits);
-#endif // if _MSC_VER
+                uint32_t offset = static_cast<uint32_t>(__builtin_clz(bits));
+#endif
 
                 // Found first bit set in bitmap
                 return item + offset;
             }
 
             // There are 32 items on each bitmap item.
-            item = item + 32UL;
+            item = item + 32u;
         }
 
         return base_;
@@ -259,8 +257,8 @@ public:
             uint32_t diff = d_func(item, base_);
             num_bits_ = std::max(diff + 1, num_bits_);
             uint32_t pos = diff >> 5;
-            diff &= 31UL;
-            bitmap_[pos] |= (1UL << (31UL - diff));
+            diff &= 31u;
+            bitmap_[pos] |= (1u << (31u - diff) );
             return true;
         }
 
@@ -300,10 +298,10 @@ public:
         num_bits_ = std::max(num_bits_, offset + n_bits);
 
         uint32_t pos = offset >> 5;             // Item position
-        offset &= 31UL;                         // Bit position inside item
+        offset &= 31u;                          // Bit position inside item
         uint32_t mask = full_mask;              // Mask with all bits set
         mask >>= offset;                        // Remove first 'offset' bits from mask
-        uint32_t bits_in_mask = 32UL - offset;  // Take note of number of set bits in mask
+        uint32_t bits_in_mask = 32u - offset;   // Take note of number of set bits in mask
 
         // This loop enters whenever the whole mask should be added
         while (n_bits >= bits_in_mask)
@@ -339,8 +337,8 @@ public:
             Diff d_func;
             uint32_t diff = d_func(item, base_);
             uint32_t pos = diff >> 5;
-            diff &= 31UL;
-            bitmap_[pos] &= ~(1UL << (31UL - diff));
+            diff &= 31u;
+            bitmap_[pos] &= ~(1u << (31u - diff));
 
             if (item == max_value)
             {
@@ -363,7 +361,7 @@ public:
             uint32_t& num_longs_used) const noexcept
     {
         num_bits = num_bits_;
-        num_longs_used = (num_bits_ + 31UL) / 32UL;
+        num_longs_used = (num_bits_ + 31u) / 32u;
         bitmap = bitmap_;
     }
 
@@ -379,9 +377,9 @@ public:
             const uint32_t* bitmap) noexcept
     {
         num_bits_ = std::min(num_bits, NBITS);
-        uint32_t num_items = ((num_bits_ + 31UL) / 32UL);
-        uint32_t num_bytes = num_items * sizeof(uint32_t);
-        bitmap_.fill(0UL);
+        uint32_t num_items = ((num_bits_ + 31u) / 32u);
+        uint32_t num_bytes = num_items * static_cast<uint32_t>(sizeof(uint32_t));
+        bitmap_.fill(0u);
         memcpy(bitmap_.data(), bitmap, num_bytes);
         if (0 < num_bits)
         {
@@ -402,7 +400,7 @@ public:
         T item = base_;
 
         // Traverse through the significant items on the bitmap
-        uint32_t n_longs = (num_bits_ + 31UL) / 32UL;
+        uint32_t n_longs = (num_bits_ + 31u) / 32u;
         for (uint32_t i = 0; i < n_longs; i++)
         {
             // Traverse through the bits set on the item, msb first.
@@ -416,21 +414,21 @@ public:
 #if _MSC_VER
                 unsigned long bit;
                 _BitScanReverse(&bit, bits);
-                uint32_t offset = 31UL ^ bit;
+                uint32_t offset = 31u ^ bit;
 #else
-                uint32_t offset = __builtin_clz(bits);
-                uint32_t bit = 31UL ^ offset;
-#endif // if _MSC_VER
+                uint32_t offset = static_cast<uint32_t>(__builtin_clz(bits));
+                uint32_t bit = 31u ^ offset;
+#endif
 
                 // Call the function for the corresponding item
                 f(item + offset);
 
                 // Clear the most significant bit
-                bits &= ~(1UL << bit);
+                bits &= ~(1u << bit);
             }
 
             // There are 32 items on each bitmap item.
-            item = item + 32UL;
+            item = item + 32u;
         }
     }
 
@@ -474,7 +472,7 @@ private:
                 // bbbbbccc bbbbbbbb cccccccc dddddddd
                 // bbbbbccc cccccddd ddddd000 dddddddd
                 // bbbbbccc cccccddd ddddd000 00000000
-                uint32_t overflow_bits = 32UL - n_bits;
+                uint32_t overflow_bits = 32u - n_bits;
                 size_t last_index = NITEMS - 1u;
                 for (size_t i = 0, n = n_items; n < last_index; i++, n++)
                 {
@@ -522,7 +520,7 @@ private:
                 // aaaaaaaa bbbbbbbb aaabbbbb bbbccccc
                 // aaaaaaaa 000aaaaa aaabbbbb bbbccccc
                 // 00000000 000aaaaa aaabbbbb bbbccccc
-                uint32_t overflow_bits = 32UL - n_bits;
+                uint32_t overflow_bits = 32u - n_bits;
                 size_t last_index = NITEMS - 1u;
                 for (size_t i = last_index, n = last_index - n_items; n > 0; i--, n--)
                 {
@@ -557,16 +555,15 @@ private:
 #if _MSC_VER
                 unsigned long bit;
                 _BitScanReverse(&bit, bits);
-                uint32_t offset = (31UL ^ bit) + 1;
+                uint32_t offset = (31u ^ bit) + 1;
 #else
-                uint32_t offset = __builtin_clz(bits) + 1;
-#endif // if _MSC_VER
-                num_bits_ = (i << 5UL) + offset;
+                uint32_t offset = static_cast<uint32_t>(__builtin_clz(bits)) + 1u;
+#endif
+                num_bits_ = (i << 5u) + offset;
                 break;
             }
         }
     }
-
 };
 
 }   // namespace fastrtps
